Fix misaligned instruction address handling
Before, Lemoncore would trap on the subsequent instruction after a jump
or branch to a misaligned address. However, this is incorrect since it
results in mepc being set to the misaligned address instead of the
address of the misaligned jump/branch.

To solve this, this commit sets misaligned_address to 1 when pc_d is
misaligned and the next state is FETCH. To prevent a logic loop, this
commit also reworks the exception logic a bit (to ensure the exception
signal does not affect ctrl_state_next directly).